trainer/pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "Trainer to train a YOLOv8 model using my Roboflow dataset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "roboflow>=1.1.63",
    "ultralytics>=8.3.131",
    "torch",
    "torchvision",
    "python-dotenv>=1.1.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[tool.ruff]
select = ["E", "W", "F", "I"]
ignore = ["E501", "W191"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
line-ending = "auto"
